chore(root): flatten directory by moving rootroot to root
