+++ internal/root/routes.go
@@ -2,10 +2,8 @@ package root
 
 import (
 	"net/http"
-
-	"cognito-repeater-go/internal/root/rootroot"
 )
 
 func RegisterRootRoutes(mux *http.ServeMux) {
-	mux.HandleFunc("/", rootroot.NewRootHandler)
+	mux.HandleFunc("/", NewRootHandler)
 }
